fallout from change, move constants into submodule
